Changes in download certificates.
start and end are int now.
To download certificates from a CT log server, the job is split into numOfWorkers,
and each one will iteratively make http requests until its portion is completed.
Added UTs.

+++ tests/benchmark/mapserver_benchmark/wholesys_benchmark/main.go
@@ -54,7 +54,7 @@ func doUpdater() {
 		wholeStart := time.Now()
 		start := time.Now()
 		err = mapUpdater.UpdateFromCT(ctx, "https://ct.googleapis.com/logs/argon2021",
-			int64(2000000+i*10000), int64(2009999+i*10000))
+			2000000+i*10000, 2009999+i*10000)
 		if err != nil {
 			panic(err)
 		}
